refactor: refactor vulnerabilities
