--- src/lib/singleview/vulnerabilities/vulnerability/Acknowledgements.svelte
@@ -1,30 +1,0 @@
-<script lang="ts">
-  import Collapsible from "$lib/Collapsible.svelte";
-  import ValueField from "$lib/singleview/ValueField.svelte";
-  import ValueList from "$lib/singleview/ValueList.svelte";
-  export let vulnerability: any;
-</script>
-
-<Collapsible header="Acknowledgements" level="4">
-  <div class="spacer" />
-  {#each vulnerability.acknowledgments as ack}
-    {#if ack.names}
-      <ValueList label="Names" values={ack.names} />
-    {/if}
-    {#if ack.organization}
-      <ValueField label="Organization" value={ack.organization} />
-    {/if}
-    {#if ack.summary}
-      <ValueField label="Summary" value={ack.summary} />
-    {/if}
-    {#if ack.urls}
-      <ValueList label="URLs" values={ack.urls} />
-    {/if}
-  {/each}
-</Collapsible>
-
-<style>
-  .spacer {
-    margin-top: 1em;
-  }
-</style>
