+++ src/lib/singleview/vulnerabilities/vulnerability/Vulnerability.svelte
@@ -13,14 +13,14 @@
   import { tick } from "svelte";
   import ProductStatus from "./ProductStatus.svelte";
   import GeneralSection from "./GeneralSection.svelte";
-  import Acknowledgements from "./Acknowledgements.svelte";
+  import Acknowledgements from "./acknowledgements/Acknowledgements.svelte";
   import ID from "./ID.svelte";
   import Notes from "./Notes.svelte";
-  import Remediations from "./Remediations.svelte";
-  import Flags from "./Flags.svelte";
-  import Involvements from "./Involvements.svelte";
-  import References from "./References.svelte";
-  import Threats from "./Threats.svelte";
+  import Remediations from "./remediations/Remediations.svelte";
+  import Flags from "./flags/Flags.svelte";
+  import Involvements from "./involvements/Involvements.svelte";
+  import References from "./references/References.svelte";
+  import Threats from "./threats/Threats.svelte";
   import Scores from "./scores/Scores.svelte";
 
   export let vulnerability: any;
