+++ src/lib/singleview/vulnerabilities/vulnerability/acknowledgements/Acknowledgement.svelte
@@ -1,0 +1,24 @@
+<script lang="ts">
+  import KeyValues from "$lib/singleview/KeyValues.svelte";
+  import ValueList from "$lib/singleview/ValueList.svelte";
+  export let ack: any;
+  const keys: Array<String> = [];
+  const values: any = [];
+  if (ack.names) {
+    keys.push("Names");
+    values.push(ack.names.join(", "));
+  }
+  if (ack.organization) {
+    keys.push("Organization");
+    values.push(ack.organization);
+  }
+  if (ack.summary) {
+    keys.push("Summary");
+    values.push(ack.summary);
+  }
+</script>
+
+<KeyValues {keys} {values} />
+{#if ack.urls}
+  <ValueList label="URLs" values={ack.urls} />
+{/if}
