+++ src/lib/singleview/vulnerabilities/vulnerability/acknowledgements/Acknowledgements.svelte
@@ -1,0 +1,18 @@
+<script lang="ts">
+  import Collapsible from "$lib/Collapsible.svelte";
+  import Acknowledgement from "./Acknowledgement.svelte";
+  export let vulnerability: any;
+</script>
+
+<Collapsible header="Acknowledgements" level="4">
+  <div class="spacer" />
+  {#each vulnerability.acknowledgments as ack}
+    <Acknowledgement {ack} />
+  {/each}
+</Collapsible>
+
+<style>
+  .spacer {
+    margin-top: 1em;
+  }
+</style>
