+++ src/lib/singleview/vulnerabilities/vulnerability/flags/Flags.svelte
