+++ src/lib/singleview/vulnerabilities/vulnerability/references/References.svelte
