+++ src/lib/singleview/vulnerabilities/vulnerability/threats/Threats.svelte
